+++ index.js
@@ -177,9 +177,38 @@ function generateSite (opt, callback) {
   callback(null)
 }
 
+/**
+ * Run sitedown and watch for changes.
+ *
+ * @param  {Object} opt   source, layout, output, silent, files, pretty
+ */
+function watch (options) {
+  var gaze = require('gaze')
+  var source = path.resolve(options.source)
+
+  sitedown(options, function (err) {
+    if (err) return console.error(err.message)
+
+    gaze('**/*.md', { cwd: source }, function (err, watcher) {
+      if (err) console.error(err.message)
+
+      console.log('\nWatching ' + source + ' for changes...')
+
+      watcher.on('all', function (event, filepath) {
+        console.log('\n' + filepath + ' was ' + event + '\n')
+
+        sitedown(options, function (err) {
+          if (err) return console.error(err.message)
+        })
+      })
+    })
+  })
+}
+
 sitedown.mdToHtml = mdToHtml
 sitedown.buildPage = buildPage
 sitedown.rewriteLinks = rewriteLinks
 sitedown.generateSite = generateSite
+sitedown.watch = watch
 
 module.exports = sitedown